backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "celery>=5.6.2",
    "cryptography>=46.0.5",
    "dj-rest-auth>=7.0.2",
    "django>=5.2.11",
    "django-allauth>=65.14.1",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.2",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "redis>=7.1.1",
    "requests>=2.32.5",
]
